(structureTensorParticleSim) Remove the last coordinate swaps in the our pipeline. ZXY is officially dead (#24). Also use doubles instead of floats.

+++ segmentation/structureTensorParticleSim/particle.cpp
@@ -1,12 +1,12 @@
 #include "particle.h"
 
-Particle::Particle(cv::Vec3f position) {
+Particle::Particle(cv::Vec3d position) {
   _position = position;
   _isStopped = false;
 }
 
 // Position in 3D space (Slice, X, Y)
-cv::Vec3f Particle::position() {
+cv::Vec3d Particle::position() {
   return _position;
 }
 
@@ -21,7 +21,7 @@ void Particle::stop() {
 }
 
 // Component wise operators
-void Particle::operator+=(cv::Vec3f v) {
+void Particle::operator+=(cv::Vec3d v) {
   _position += v;
 }
 
@@ -29,6 +29,6 @@ float Particle::operator()(int index) {
   return _position(index);
 }
 
-cv::Vec3f Particle::operator-(Particle p) {
+cv::Vec3d Particle::operator-(Particle p) {
   return _position - p.position();
 }
